test(ConnectionProvider): 연결 테스트
