+++ src/test/java/chess/controller/CommandLineTest.java
@@ -1,15 +1,14 @@
-package chess;
+package chess.controller;
 
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-import chess.controller.CommandLine;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 class CommandLineTest {
-    
+
     @Test
     @DisplayName("명령 정상 생성 테스트")
     void create_command_test() {
@@ -17,7 +16,7 @@ class CommandLineTest {
         assertDoesNotThrow(() -> new CommandLine(List.of("end")));
         assertDoesNotThrow(() -> new CommandLine(List.of("move", "a1", "b2")));
     }
-    
+
     @Test
     @DisplayName("명령 생성 에러 테스트")
     void create_error_command() {
