+++ src/test/java/chess/dao/ConnectionProviderTest.java
@@ -1,0 +1,18 @@
+package chess.dao;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.*;
+
+import java.sql.SQLException;
+import org.junit.jupiter.api.Test;
+
+class ConnectionProviderTest {
+    @Test
+    public void connection() {
+        try (final var connection = ConnectionProvider.getConnection()) {
+            assertThat(connection).isNotNull();
+        } catch (SQLException e) {
+            throw new RuntimeException(e);
+        }
+    }
+}
